database/pyproject.toml
[project]
name = "database-test"
version = "0.1.0"
description = "Project to test the database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "click>=8.2.1",
    "load-dotenv>=0.1.0",
    "supabase>=2.18.0",
]
